Change boost::posix_time::milliseconds init to int64_t (#105)
* Change boost::posix_time::milliseconds init to uint64_t

Fixes #104

* Update simple_action_server_imp.h

* Update destruction_guard.h

* Update simple_action_client.h

* use int64_t

+++ include/actionlib/client/simple_action_client.h
@@ -601,7 +601,8 @@ bool SimpleActionClient<ActionSpec>::waitForResult(const ros::Duration & timeout
       time_left = loop_period;
     }
 
-    done_condition_.timed_wait(lock, boost::posix_time::milliseconds(time_left.toSec() * 1000.0f));
+    done_condition_.timed_wait(lock,
+      boost::posix_time::milliseconds(static_cast<int64_t>(time_left.toSec() * 1000.0f)));
   }
 
   return cur_simple_state_ == SimpleGoalState::DONE;
